+++ backend/src/routes/odps/odp.module.ts
@@ -1,5 +1,6 @@
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
+import { OdpStore } from './odp-store.service';
 import { OdpController } from './odp.controller';
 import { OdpService } from './odp.service';
 
@@ -12,7 +13,8 @@ import { OdpService } from './odp.service';
 		OdpController
 	],
 	providers: [
-		OdpService
+		OdpService,
+		OdpStore
 	],
 })
 export class OdpModule {}
